TP_2/requirements.txt
beautifulsoup4 
lxml
Pillow
selenium
aiofiles
aiohttp
requests